+++ JobOverview/Entities/Travail.cs
@@ -4,7 +4,7 @@
     {
         public DateOnly DateTravail { get; set; }
         public int IdTache { get; set; }
-        public float Heures { get; set; }
+        public decimal Heures { get; set; }
         public float TauxProductivite { get; set; }
     }
 }
